Working on generic item containing slot
